Added Vehicle Smoke, Particle Effect & Edited the BasicEnemy
I incorporated the Vehicle Smoke Sprite Sheet PNG from Valentine and converted it into a Sprite Sheet in Unity.

Next, I established a Particle System named 'Vehicle Smoke' to emit the Sprite Sheet.

However, the Particle System required a Material. Therefore, I generated a new Material named 'VehicleSmokeMaterial' and assigned 'VehicleSmokeSpriteSheet' as its material.

Once the process was completed, I created a 'VehicleSmoke-Prefab' for convenient usage with other vehicles.

Additionally, I configured the 'BasicEnemy' Prefab to utilize the VehicleSmoke.

The particle effects still looks a bit off tho, as it wouldn't rotate with the car.
I don't know a Solution for this, so we'll have to see how to fix it.